+++ packages/react-native/src/NotifeeApiModule.ts
@@ -16,6 +16,7 @@ import {
   Event,
   TriggerNotification,
   DisplayedNotification,
+  NotificationSettings,
 } from './types/Notification';
 import { PowerManagerInfo } from './types/PowerManagerInfo';
 import { Trigger } from './types/Trigger';
@@ -37,11 +38,7 @@ import validateNotification from './validators/validateNotification';
 import validateTrigger from './validators/validateTrigger';
 import validateAndroidChannel from './validators/validateAndroidChannel';
 import validateAndroidChannelGroup from './validators/validateAndroidChannelGroup';
-import {
-  IOSNotificationCategory,
-  IOSNotificationSettings,
-  IOSNotificationPermissions,
-} from './types/NotificationIOS';
+import { IOSNotificationCategory, IOSNotificationPermissions } from './types/NotificationIOS';
 import validateIOSCategory from './validators/validateIOSCategory';
 import validateIOSPermissions from './validators/validateIOSPermissions';
 
@@ -460,24 +457,28 @@ export default class NotifeeApiModule extends NotifeeNativeModule implements Mod
 
   public requestPermission = (
     permissions: IOSNotificationPermissions = {},
-  ): Promise<IOSNotificationSettings> => {
+  ): Promise<NotificationSettings> => {
     if (isAndroid) {
-      // Android doesn't require permission, so instead we
-      // return a dummy response to allow the permissions
-      // flow work the same on both iOS & Android
-      return Promise.resolve({
-        alert: 1,
-        badge: 1,
-        criticalAlert: 1,
-        showPreviews: 1,
-        sound: 1,
-        carPlay: 1,
-        lockScreen: 1,
-        announcement: 1,
-        notificationCenter: 1,
-        inAppNotificationSettings: 1,
-        authorizationStatus: 1,
-      } as IOSNotificationSettings);
+      return this.native
+        .getNotificationSettings()
+        .then(({ authorizationStatus }: Pick<NotificationSettings, 'authorizationStatus'>) => {
+          return {
+            authorizationStatus,
+            ios: {
+              alert: 1,
+              badge: 1,
+              criticalAlert: 1,
+              showPreviews: 1,
+              sound: 1,
+              carPlay: 1,
+              lockScreen: 1,
+              announcement: 1,
+              notificationCenter: 1,
+              inAppNotificationSettings: 1,
+              authorizationStatus,
+            },
+          };
+        });
     }
 
     let options: IOSNotificationPermissions;
@@ -535,24 +536,28 @@ export default class NotifeeApiModule extends NotifeeNativeModule implements Mod
     return this.native.getNotificationCategories();
   };
 
-  public getNotificationSettings = (): Promise<IOSNotificationSettings> => {
+  public getNotificationSettings = (): Promise<NotificationSettings> => {
     if (isAndroid) {
-      // Android doesn't support this, so instead we
-      // return a dummy response to allow the permissions
-      // flow work the same on both iOS & Android
-      return Promise.resolve({
-        alert: 1,
-        badge: 1,
-        criticalAlert: 1,
-        showPreviews: 1,
-        sound: 1,
-        carPlay: 1,
-        lockScreen: 1,
-        announcement: 1,
-        notificationCenter: 1,
-        inAppNotificationSettings: 1,
-        authorizationStatus: 1,
-      } as IOSNotificationSettings);
+      return this.native
+        .getNotificationSettings()
+        .then(({ authorizationStatus }: Pick<NotificationSettings, 'authorizationStatus'>) => {
+          return {
+            authorizationStatus,
+            ios: {
+              alert: 1,
+              badge: 1,
+              criticalAlert: 1,
+              showPreviews: 1,
+              sound: 1,
+              carPlay: 1,
+              lockScreen: 1,
+              announcement: 1,
+              notificationCenter: 1,
+              inAppNotificationSettings: 1,
+              authorizationStatus,
+            },
+          };
+        });
     }
 
     return this.native.getNotificationSettings();